Revert "refactor MediaService"
This reverts commit 7aa325f914e0bf47e4a575545d9f547645188bb2.
